# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SimAPI",
    "description": "The CI/CD layer for engineering simulations. Dual-layer validation: 280+ deterministic physics checks across 21 domains plus optional LLM reasoning.",
    "version": "3.1.0"
  },
  "paths": {
    "/v1/health": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Health",
        "description": "Liveness + basic service facts. Unauthenticated by design.",
        "operationId": "health_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health V1 Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/metrics": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Prometheus Metrics",
        "description": "Prometheus text-format metrics for scraping.",
        "operationId": "prometheus_metrics_v1_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/validate": {
      "post": {
        "tags": [
          "validation"
        ],
        "summary": "Validate",
        "operationId": "validate_v1_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/validate/upload": {
      "post": {
        "tags": [
          "validation"
        ],
        "summary": "Validate Upload",
        "operationId": "validate_upload_v1_validate_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_upload_v1_validate_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/validate/physics-only": {
      "post": {
        "tags": [
          "validation"
        ],
        "summary": "Validate Physics Only",
        "operationId": "validate_physics_only_v1_validate_physics_only_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/job/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_v1_job__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/job/{job_id}/ai": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Ai",
        "description": "Poll for the AI result once it is ready.",
        "operationId": "get_job_ai_v1_job__job_id__ai_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "description": "List recent jobs, newest first, with cursor-free offset pagination.",
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page size.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of jobs to skip.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of jobs to skip."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/demo": {
      "post": {
        "tags": [
          "validation"
        ],
        "summary": "Demo",
        "description": "Run a validation against synthetic aerodynamics data with seeded anomalies.",
        "operationId": "demo_v1_demo_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_validate_upload_v1_validate_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "simulation_type": {
            "type": "string",
            "title": "Simulation Type",
            "default": "aerodynamics"
          },
          "conditions": {
            "type": "string",
            "title": "Conditions",
            "default": "{}"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "default": ""
          },
          "run_ai": {
            "type": "string",
            "title": "Run Ai",
            "default": "true"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_validate_upload_v1_validate_upload_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SimulationType": {
        "type": "string",
        "enum": [
          "aerodynamics",
          "fluid_dynamics",
          "structural",
          "thermodynamics",
          "robotics",
          "combustion",
          "acoustics",
          "electromagnetics",
          "geomechanics",
          "biomechanics",
          "nuclear",
          "plasma",
          "chemical",
          "hydrodynamics",
          "meteorology",
          "astrophysics",
          "materials",
          "tribology",
          "aeroelasticity",
          "cryogenics",
          "multiphysics"
        ],
        "title": "SimulationType"
      },
      "ValidateRequest": {
        "properties": {
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Data",
            "description": "Trials as a list of records."
          },
          "simulation_type": {
            "$ref": "#/components/schemas/SimulationType",
            "description": "Physics domain to validate against.",
            "default": "aerodynamics"
          },
          "conditions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Conditions",
            "description": "Input boundary conditions."
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id",
            "description": "Optional caller-supplied tracking id."
          },
          "run_ai": {
            "type": "boolean",
            "title": "Run Ai",
            "description": "Run the async AI reasoning layer.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ValidateRequest",
        "description": "Payload for JSON validation requests."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}